.ci/lumen_cli/pyproject.toml
[project]
name = "lumen"
version = "0.1.0"
dependencies = [
    "pyyaml==6.0.2",
    "GitPython==3.1.45",
    "docker==7.1.0",
    "pytest==7.3.2",
    "uv==0.9.6"
]

[project.scripts]
lumen = "cli.run:main"

[tool.setuptools]
packages = ["cli"]

[tool.setuptools.package-dir]
cli = "cli"

[tool.ruff.lint]
# Enable preview mode for linting
preview = true

# Now you can select your preview rules, like RUF048
extend-select = ["RUF048"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]
